исправление int

+++ untitled/src/BonusMilesService.java
@@ -1,5 +1,5 @@
 public class BonusMilesService {
-    public long calculate(int cost) {
+    public int calculate(int cost) {
         int rubleMile = 20;
         int Miles = cost / rubleMile;
